+++ full-backup.sh
@@ -161,7 +161,7 @@ if [ ! -z "$PASSPHRASE" ]; then
     else
         echo; echo 'Backup successfully encrypted'; echo
     fi
-    rm $LATEST_FULL_BACKUP_FILE_NAME
+    [ -f $LATEST_FULL_BACKUP_FILE_NAME ] && rm $LATEST_FULL_BACKUP_FILE_NAME
     ln -s ${ENCRYPTED_FULL_BACKUP_FILE_NAME} $LATEST_FULL_BACKUP_FILE_NAME
 else
     # let's do it using tar
@@ -179,7 +179,7 @@ else
     echo "[WARNING]"
     echo "If your data came from Europe, please check GDPR compliance."
 
-    rm $LATEST_FULL_BACKUP_FILE_NAME
+    [ -f $LATEST_FULL_BACKUP_FILE_NAME ] && rm $LATEST_FULL_BACKUP_FILE_NAME
     ln -s ${FULL_BACKUP_FILE_NAME} $LATEST_FULL_BACKUP_FILE_NAME
 fi
